Checkin is now a Cog and out of Utils
